+++ src/index.js
@@ -1,51 +1,14 @@
 import './style.css';
+import sendUser from './scores.js';
+import getData from './data.js';
 
 const list = document.querySelector('.container');
 const refresh = document.querySelector('.refresh');
 
 
-const nameField = document.querySelector('.name');
-const scoreField = document.querySelector('.scores');
+const name = document.querySelector('.name');
+const score = document.querySelector('.scores');
 const submit = document.querySelector('.btn')
-
-// const sendData = async ()=>{
-//   const data = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games',{
-//       method: 'POST',
-//       headers: {
-//         'Accept': 'application/json',
-//         'Content-Type': 'application/json'
-//       },
-//       body: JSON.stringify({"name": "Football" })
-//     });
-
-//     const content = await data.json();
-//     console.log(content);
-// }
-
-// sendData()
-
-// send the user and scores 
-const sendScores = async ()=>{
-  const data = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/a693dn4AamOZ5GxAOLyY/scores/',{
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({"user":nameField.value, "score":scoreField.value})
-    });
-    const content = await data.json();
-    console.log(content);
-}
-
-//Get Data 
-const getData = async(callback)=>{
-  const url = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/a693dn4AamOZ5GxAOLyY/scores/'
-const data = await fetch(url);
-const content = await data.json();
-console.log(content.result)
-callback(content.result)
-}
 
 // Display Data 
 const display = scores=>{
@@ -58,9 +21,9 @@ const display = scores=>{
 
 submit.addEventListener('click',(e)=>{
   e.preventDefault();
-  sendScores();
-  nameField.value="";
-  scoreField.value="";
+  sendUser(name.value,score.value);
+  name.value="";
+  score.value="";
 });
 
 // refresh
